0125 in class

+++ My Little Witch/Assets/03.Scripts/Data/MonData.cs
@@ -15,7 +15,7 @@ public class MonData : ScriptableObject
     public float attackRadius;
     public float attackSpeed;
     public Vector3 HPlocalScale;
-    public float EXP;
+    public int EXP;
 
     public ItemData[] DropItems;
 }
